Add pgactive.connectability_check_duration GUC
This new "internal" GUC sets the total amount of time the per-db worker should
try to connect in case of failed attempts. On some configuration, during the
engine startup, this worker can be spawned too early and not be able to connect
yet. The duration between each attempt is 1 second.