Convey per-db worker intention of unregister to supervisor (#275)
Currently, per-db worker when it unregisters tells postgres to not
restart (via proc_exit(0)), but doesn't tell the supervisor to not
re-register. Due to this, when the supervisor next wakes up and
does the db scan, it registers the per-db worker again and
postgres starts a new one. This approach doesn't honour the per-db
worker's intention to unregister.

As a side effect of this kind of unregistering, the supervisor,
due to a race condition, gets stuck in loop in
pgactive_register_perdb_worker() when a per-db worker unregisters
immediately after it is started, e.g. when local connectability
check fails.

To fix these issues, the per-db worker unregistering logic is
changed to the following way:
- When unregistered, a flag is set in per-db worker slot in
  pgactive shared memory without releaseing the slot. This flag
  can be seen via pgactive_get_workers_info() SQL function.
- When found a previously unregistered per-db worker, the
  supervisor skips registering the per-db worker again.

Since the pgactive shared memory slot for unregistered per-db
worker is not released, the worker slots might get consumed if
there are a lot of unregistered per-db workers. This might be okay
because the chances of per-db workers unregistering is infrequent.