Fix issues with pgactive replication lag timestamps reporting
When the xact_committs are 0, pgactive_node_slots reports them as
the epoch time because TimestampTzGetDatum(0) gives epoch time.
So, fix it by reporting NULL if any of xact_committs are 0.

Also, correct the pgactive_node_slots example in the docs which
shows the diff between sent and applied xact_committs and xid as
lag. The diff can get negative because the order of sending
logical decoding changes and applying the changes are in the
transaction commit order, so it can't be assumed that sent xid is
always <= applied xid. One must treat these xact_committs and xid
as additional replication lag information.

Also, fix the CONTEXT message in apply side which can have
TimestampTzGetDatum(0) printed as epoch when the replication
origin XID, LSN and committs are set to 0 while the context
message callback is in effect.